+++ src/export.ts
@@ -2,15 +2,17 @@ import Button from './lib/button/button';
 import Chip from './lib/chip';
 import DropDownMenu from './lib/menu/dropdown_menu';
 import FileInput from './lib/form/file_input';
+import FormDialog, { FormDialogButton as FormDialogButtonType } from './lib/dialog/formDialog';
 import IconButton from './lib/button/icon_button';
 import Input from './lib/form/input';
 import InputSet from './lib/form/input_set';
 import ItemRow from './lib/list/item_row';
 import ItemRowSearch from './lib/list/item_row_search';
 import List, { ListItem as ListItemType } from './lib/list/list';
-import Modal from './lib/modal';
+import Modal from './lib/dialog/modal';
 import PDFViewer from './lib/pdf_viewer/pdf_viewer';
 import PeriodSelector, { PeriodOptions } from './lib/date_time/period_selector';
+import Table, { TableSortDirection as TableSortDirectionType } from './lib/table/table';
 import Textarea from './lib/form/textarea';
 import TextareaSet from './lib/form/textarea_set';
 import TimePicker from './lib/form/time_picker';
@@ -24,6 +26,7 @@ export {
     Chip,
     DropDownMenu,
     FileInput,
+    FormDialog,
     IconButton,
     Input,
     InputSet,
@@ -34,6 +37,7 @@ export {
     PDFViewer,
     PeriodOptions,
     PeriodSelector,
+    Table,
     Textarea,
     TextareaSet,
     TimePicker,
@@ -44,4 +48,6 @@ export {
 
 
 export type ListItem = ListItemType
+export type FormDialogButton = FormDialogButtonType
+export type TableSortDirection = TableSortDirectionType