+++ src/lib/pdf_viewer/pdf_renderer.ts
@@ -56,7 +56,7 @@ export async function renderPDFPageAsCanvas(
 	}	
 	let viewport = page.getViewport({
 		scale: scalingFactor + zoom, 
-		rotation: rotation 
+		rotation: page.rotate === 0 ? rotation : rotation - page.rotate,
 	});	
 	canvas.height = viewport.height;
 	canvas.width = viewport.width;