+++ src/lib/private/types.tsx
@@ -1,0 +1,1 @@
+export type ColorOptionType = 'primary' | 'secondary' | 'tertiary' | 'error' | 'success' | 'alert' | 'accessory-blue' | 'accessory-orange' | 'accessory-green';